[VAS-541] Refactored API methods
