+++ src/pages/stations/stationAssociateEC/StationAssociateECPage.tsx
@@ -23,7 +23,7 @@ import {
     CreditorInstitutionAssociatedCodeList
 } from '../../../api/generated/portal/CreditorInstitutionAssociatedCodeList';
 import {Delegation} from '../../../api/generated/portal/Delegation';
-import {getBrokerDelegation} from "../../../services/channelService";
+import { getBrokerDelegation } from '../../../services/institutionService';
 import {ProblemJson} from '../../../api/generated/portal/ProblemJson';
 import {isErrorResponse} from '../../../utils/client-utils';
 import ECSelectionSearch from './ECSelectionSearch';
