+++ src/services/__mocks__/channelService.ts
@@ -12,11 +12,12 @@ import {
   StatusEnum,
   WrapperChannelDetailsDto,
 } from '../../api/generated/portal/WrapperChannelDetailsDto';
-import { TypeEnum, WrapperEntities } from '../../api/generated/portal/WrapperEntities';
+import { WrapperEntities } from '../../api/generated/portal/WrapperEntities';
 import { ChannelOnCreation } from '../../model/Channel';
 import { WfespPluginConfs } from '../../api/generated/portal/WfespPluginConfs';
 import { Delegation } from '../../api/generated/portal/Delegation';
 import { mockedPaymentTypes } from './configurationService';
+import { channelWrapperMockedGet } from './institutionsService';
 
 export const mockedChannels: ChannelsResource = {
   channels: [
@@ -389,41 +390,6 @@ export const channelCode: ChannelCodeResource = {
   channel_code: '1231231231',
 };
 
-export const channelWrapperMockedGet = (code: string): WrapperEntities => ({
-  brokerCode: 'string',
-  createdAt: new Date(),
-  createdBy: 'PSP S.p.A',
-  id: 'string',
-  modifiedAt: new Date(),
-  modifiedBy: 'string',
-  modifiedByOpt: 'string',
-  note: 'string',
-  status: StatusEnum.APPROVED,
-  type: TypeEnum.CHANNEL,
-  entities: [
-    {
-      createdAt: new Date(),
-      entity: {
-        broker_psp_code: '97735020584',
-        broker_description: 'AgID - Agenzia per l’Italia Digitale',
-        channel_code: code,
-        target_path: ' /govpay/api/pagopa/PagamentiTelematiciCCPservice',
-        target_port: 8081,
-        target_host: ' lab.link.it',
-        payment_types: mockedPaymentTypes?.payment_types?.map((e) => e.payment_type ?? "") ?? [],
-        status: StatusEnum.TO_CHECK,
-      },
-      id: 'string',
-      modifiedAt: new Date(),
-      modifiedBy: 'string',
-      modifiedByOpt: 'Operatore PSP',
-      note: 'string',
-      status: StatusEnum.TO_CHECK,
-      type: TypeEnum.CHANNEL,
-    },
-  ],
-});
-
 export const getChannels = (_page: number): Promise<ChannelsResource> =>
   new Promise((resolve) => resolve(mockedChannels));
 
@@ -461,9 +427,6 @@ export const updateChannel = (
 
 export const getChannelPSPs = (page: number): Promise<ChannelPspListResource> =>
   new Promise((resolve) => resolve(page === 0 ? mockedChannelPSPs : mockedChannelPSPsPage2));
-
-export const getBrokerDelegation = (): Promise<Array<Delegation>> =>
-  new Promise((resolve) => resolve(mockedDelegatedPSP));
 
 export const associatePSPtoChannel = (
   _channelcode: string,
